fix(consensus): harden arithmetic and eliminate unwrap in block path
- safety.rs: widen quorum calculation to u64 before multiply to prevent
  u32 overflow with large validator counts (matches protocol.rs pattern)
- validate.rs: use checked_add for nonce successor to reject u64::MAX
  overflow instead of silently wrapping
- chain.rs: replace all blocks.last().unwrap() in produce_block with
  indexed access via tip_idx, eliminating panic risk under refactoring

+++ crates/sccgub-consensus/src/safety.rs
@@ -34,8 +34,8 @@ pub struct SafetyCertificate {
 impl SafetyCertificate {
     /// Verify that the certificate is structurally valid (quorum size, no duplicates).
     pub fn verify_structure(&self) -> Result<(), String> {
-        let expected_quorum = (2 * self.validator_count) / 3 + 1;
-        if self.quorum != expected_quorum {
+        let expected_quorum = (2u64 * self.validator_count as u64) / 3 + 1;
+        if self.quorum as u64 != expected_quorum {
             return Err(format!(
                 "Quorum mismatch: claimed {} but expected {} for n={}",
                 self.quorum, expected_quorum, self.validator_count
@@ -116,7 +116,7 @@ impl SafetyCertificate {
         precommits: &HashMap<Hash, crate::protocol::Vote>,
         validator_count: u32,
     ) -> Self {
-        let quorum = (2 * validator_count) / 3 + 1;
+        let quorum = ((2u64 * validator_count as u64) / 3 + 1) as u32;
         let precommit_signatures: Vec<(Hash, Vec<u8>)> = precommits
             .values()
             .filter(|v| v.block_hash == block_hash)
